requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.3
djangorestframework==3.15.2
greenlet==3.1.1
gunicorn==23.0.0
numpy==2.1.3
packaging==24.2
pandas==2.2.3
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
SQLAlchemy==2.0.36
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
